refactor: split chat commands
- extract command handlers into dedicated modules

- add shared types/utils for command processing

- update dispatcher imports and command routing
