+++ apps/remote-server/src/core/chat-commands/help.ts
@@ -1,0 +1,48 @@
+export function buildHelpMessage(): string {
+  return [
+    '📋 可用命令：',
+    '/help - 查看命令帮助',
+    '/ping - 检查机器人在线状态',
+    '/new - 创建新会话',
+    '/restart - 直接执行 pm2 restart（可用 PM2_PROCESS_NAME 覆盖进程名）',
+    '/restart update [branch] - 切到分支拉取最新代码，build 后再重启（默认 master）',
+    '/restart status - 查看最近一次重启结果',
+    '/clear - 清空当前会话上下文',
+    '/reset - /clear 的别名',
+    '/compact - 强制压缩当前会话上下文',
+    '/mode - 查看当前模式',
+    '/mode planning - 切换到 planning 模式',
+    '/mode executing - 切换到 executing 模式',
+    '/model - 查看当前模型',
+    '/model reset - 恢复默认模型',
+    '/model <name> - 切换模型（写入 config.json）',
+    '/memory - 查看 memory 列表',
+    '/memory on|off - 开关当前会话 memory',
+    '/memory pin <id> - 置顶一条 memory',
+    '/memory forget <id> - 删除一条 memory',
+    '/current - 查看当前绑定会话',
+    '/detach - 断开当前会话绑定（不删除历史）',
+    '/resume - 查看历史会话（最近 10 条）',
+    '/resume [list] [N] - 查看最近 N 条历史会话（N 范围 1-30）',
+    '/sessions - /resume 的别名',
+    '/resume <session-id> - 恢复指定会话',
+    '/fork <session-id> - 基于指定会话创建新分叉会话并自动切换',
+    '/status - 查看当前运行状态与会话信息',
+    '/stop - 停止当前正在运行的任务',
+    '/cancel - /stop 的别名',
+    '/insight [days] - 汇总近 N 天会话洞见（默认 7 天，N 范围 1-365）',
+    '/skills - 查看可用技能',
+    '/skills <name|index> <message> - 使用技能',
+    '/soul - 查看可用 soul',
+    '/soul list - 列出可用 soul',
+    '/soul status - 查看当前会话 soul 状态',
+    '/soul use <id> - 切换到指定 soul',
+    '/soul add <id> - 追加一个 soul',
+    '/soul remove <id> - 移除一个 soul',
+    '/soul clear - 清除当前会话 soul',
+    '/wt status - 查看当前会话绑定的 worktree',
+    '/wt use <id> - 复用已有或创建新的 worktree 并绑定',
+    '/wt use <id> <repoRoot> <worktreePath> [branch] - 绑定/更新当前会话 worktree',
+    '/wt clear - 清除当前会话 worktree 绑定',
+  ].join('\n');
+}
